requirements.txt
fastapi==0.115.12
uvicorn
gunicorn
timescaledb
sqlmodel
pydantic
sqlalchemy
requests
python-decouple
psycopg[binary]
python-multipart
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
# psycopg2-binary